+++ src/float/mod.rs
@@ -25,6 +25,8 @@ pub(crate) mod arith;
 pub(crate) mod big;
 mod casts;
 mod cmp;
+#[cfg(feature = "num-traits")]
+mod impl_num_traits;
 mod ord;
 #[cfg(feature = "serde")]
 mod serde;