+++ src/integer/mod.rs
@@ -42,6 +42,8 @@ pub(crate) mod big;
 mod casts;
 mod cmp;
 mod division;
+#[cfg(feature = "num-traits")]
+mod impl_num_traits;
 #[cfg(feature = "serde")]
 mod serde;
 pub(crate) mod small;